+++ packages/config/src/chains.ts
@@ -47,7 +47,6 @@ import {
   ring,
   rmrk,
   sdn,
-  soon,
   sub,
   teer,
   tt1,
@@ -255,7 +254,7 @@ export const calamari = new Parachain({
   name: 'Calamari',
   parachainId: 2084,
   ss58Format: 78,
-  ws: 'wss://ws.calamari.systems',
+  ws: 'wss://calamari.systems',
 });
 
 export const centrifuge = new Parachain({
@@ -401,33 +400,6 @@ export const equilibrium = new Parachain({
   parachainId: 2011,
   ss58Format: 68,
   ws: 'wss://node.pol.equilibrium.io',
-});
-
-export const equilibriumAlphanet = new Parachain({
-  assetsData: [
-    {
-      asset: dev,
-      decimals: 9,
-      id: 1_735_159_154,
-    },
-    {
-      asset: eq,
-      id: 25_969,
-    },
-    {
-      asset: eqd,
-      id: 6_648_164,
-    },
-  ],
-  ecosystem: Ecosystem.AlphanetRelay,
-  genesisHash:
-    '0x0ba4508078664cfd04f00bf17892dd1a0764ce61d90857f76c6a2ea61f008247',
-  isTestChain: true,
-  key: 'equilibrium-alphanet',
-  name: 'EquilibriumAlphanet',
-  parachainId: 2011,
-  ss58Format: 68,
-  ws: 'wss://devnet.equilab.io/alphanet/collator/api/wss',
 });
 
 export const hydraDX = new Parachain({
@@ -625,24 +597,6 @@ export const kusamaAssetHub = new Parachain({
   ws: 'wss://kusama-asset-hub-rpc.polkadot.io',
 });
 
-export const litentryAlphanet = new Parachain({
-  assetsData: [
-    {
-      asset: lit,
-      id: 'SelfReserve',
-    },
-  ],
-  ecosystem: Ecosystem.AlphanetRelay,
-  genesisHash:
-    '0x4decfd421755d9008781d00ce40ea4cf47ec24a3717a34ca995126fc4a78c4f8',
-  isTestChain: true,
-  key: 'litentry-alphanet',
-  name: 'Litentry Alphanet',
-  parachainId: 2106,
-  ss58Format: 131,
-  ws: 'wss://moonbase-parachain-sg-0.litentry.io',
-});
-
 export const litmus = new Parachain({
   assetsData: [
     {
@@ -696,23 +650,15 @@ export const moonbaseAlpha = new EvmParachain({
       asset: atom,
       id: '244316754493307480955066032215622931381',
     },
-    {
-      asset: auq,
-      id: '69536036667157951501899290870203586130',
-    },
+    // {
+    //   asset: auq,
+    //   id: '69536036667157951501899290870203586130',
+    // },
     {
       asset: dev,
       id: '0x0000000000000000000000000000000000000802',
       metadataId: 0,
       min: 0.01,
-    },
-    {
-      asset: eq,
-      id: '190590555344745888270686124937537713878',
-    },
-    {
-      asset: eqd,
-      id: '187224307232923873519830480073807488153',
     },
     {
       asset: lit,
@@ -729,10 +675,6 @@ export const moonbaseAlpha = new EvmParachain({
     {
       asset: pica,
       id: '22417088946346045371238623691600461855',
-    },
-    {
-      asset: soon,
-      id: '205153165378836428058230526014907639736',
     },
     {
       asset: tt1,
@@ -1132,7 +1074,7 @@ export const parallel = new Parachain({
   name: 'Parallel',
   parachainId: 2012,
   ss58Format: 172,
-  ws: 'wss://rpc.parallel.fi',
+  ws: 'wss://parallel-rpc.dwellir.com',
 });
 
 export const parallelHeiko = new Parachain({
@@ -1459,7 +1401,6 @@ export const chainsList: AnyChain[] = [
   darwiniaCrab,
   darwiniaPangoro,
   equilibrium,
-  equilibriumAlphanet,
   hydraDX,
   integritee,
   interlay,
@@ -1467,7 +1408,6 @@ export const chainsList: AnyChain[] = [
   khala,
   kintsugi,
   kusama,
-  litentryAlphanet,
   litmus,
   mangataKusama,
   moonbaseAlpha,