+++ packages/config/src/configs/index.ts
@@ -13,7 +13,6 @@ import { darwiniaConfig } from './darwinia';
 import { darwiniaCrabConfig } from './darwiniaCrab';
 import { darwiniaPangoroConfig } from './darwiniaPangoro';
 import { equilibriumConfig } from './equilibrium';
-import { equilibriumAlphanetConfig } from './equilibriumAlphanet';
 import { hydraDxConfig } from './hydraDX';
 import { integriteeConfig } from './integritee';
 import { interlayConfig } from './interlay';
@@ -22,7 +21,6 @@ import { khalaConfig } from './khala';
 import { kintsugiConfig } from './kintsugi';
 import { kusamaConfig } from './kusama';
 import { kusamaAssetHubConfig } from './kusamaAssetHub';
-import { litentryAlphanetConfig } from './litentryAlphanet';
 import { litmusConfig } from './litmus';
 import { mangataKusamaConfig } from './mangataKusama';
 import { moonbaseAlphaConfig } from './moonbaseAlpha';
@@ -61,7 +59,6 @@ export const chainsConfigList: ChainConfig[] = [
   darwiniaConfig,
   darwiniaCrabConfig,
   darwiniaPangoroConfig,
-  equilibriumAlphanetConfig,
   equilibriumConfig,
   hydraDxConfig,
   integriteeConfig,
@@ -70,7 +67,6 @@ export const chainsConfigList: ChainConfig[] = [
   khalaConfig,
   kintsugiConfig,
   kusamaConfig,
-  litentryAlphanetConfig,
   litmusConfig,
   mangataKusamaConfig,
   moonbaseAlphaConfig,